Encode creation time into priority for FIFO within same level
Uses bigint priority = level * 10_000_000_000 + epoch_seconds so a
single column sort handles both type priority and creation order.

+++ db/analytics_migrate/20260227165150_add_priority_to_strava_requests.rb
@@ -1,5 +1,5 @@
 class AddPriorityToStravaRequests < ActiveRecord::Migration[6.1]
   def change
-    add_column :strava_requests, :priority, :integer, null: false, default: 0
+    add_column :strava_requests, :priority, :bigint, null: false, default: 0
   end
 end
